Updating some of it.
As some of you have tested this command. You may also know that the previous one doesn't works for animated stickers when the gif/video is of high quality, so in this one I have added options for the `Sticker Types` and `Quality` (copied from the `Sticker` command) so that it won't have any problem in formatting animated stickers. I have also fixed some typos. And as the command name indicates this command is used for stealing stickers, but when the sender doesn't gives the new author of the sticker, the author will be `Kaoi` , so I think this is not a good idea so I have changed it to the sender username. And I have also made the bot to send an example for this command when the user doesn't gives us the new pack of the sticker.